Backend/requirements.txt
blinker==1.9.0
click==8.1.8
dlib==19.24.6
dnspython==2.7.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pymongo==4.11.3
python-dotenv==1.0.1
setuptools==76.1.0
Werkzeug==3.1.3
wheel==0.45.1
